perf: optimize use of `ClangParams` struct
This clones a reference (pointer) to the same `ClangParams` instance instead of cloning the instance data (when processing each source file).

Some param types updated to reflect the actual life cycle of the `ClangParams` instance. Another param type change relies on internal mutability when passing around references to `FileObj` instances in an array.

Added auto-trait `Debug` to `ClangVersions` struct, in case debug printing is ever needed.

Removes a duplicate condition where ranges of changed lines are an empty list if `lines-changed-only` is `Off`.