fix: use diagnostic name by default (#236)
Corresponds to cpp-linter/cpp-linter#173. This should never happen in
the wild, but just in case there's an underlying problem with parsing
clang-tidy output...

This fixes a problem about hyperlinking a clang-tidy diagnostic name to
the clang-tidy docs.

Specifically when the diagnostic name does not satisfy the following
conditions:
- starts with "clang-analyzer-"
- starts with "clang-diagnostic-"
- contains at least 1 hyphen (`-`)

This also removes the `.unwrap()` calls (bad practice) and adds a test
for code coverage.